Pack Attack zone: rename zone from Packmaster to Pack Attack
