+++ src/main/resources/anniv6Resources/localization/eng/PackAttack/Eventstrings.json
