+++ src/main/resources/anniv6Resources/localization/eng/PackAttack/UIstrings.json
@@ -1,0 +1,9 @@
+{
+  "${ModID}:PackAttack": {
+    "TEXT": [
+      "Pack Attack",
+      "Cards from Packmaster packs are everywhere! Each turn, you draw one less card, but add a random card from a set of three packs to your hand. NL NL Card rewards and shops contain cards from the packs.",
+      "Cards Packmaster packs {0}, {1}, and {2} are everywhere! Each turn, you draw one less card, but add a random card from those packs to your hand. NL NL Card rewards and shops contain cards from those packs."
+    ]
+  }
+}
